python/requirements.txt
MT5Manager == 5.0.4732
python-dotenv == 1.0.1
